refactor: basic changes in the structure of some components
