+++ personal-portfolio/src/App.css
@@ -243,8 +243,10 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
   font-weight: 700;
   letter-spacing: 0.8px;
   padding: 8px 10px;
-  background: linear-gradient(90.21deg, rgba(170, 54, 124, 0.5) -5.91%, rgba(74, 47, 189, 0.5) 111.58%);
+  background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%);
   border: 1px solid rgba(255, 255, 255, 0.5);
+  /* background: linear-gradient(90.21deg, rgba(170, 54, 124, 0.5) -5.91%, rgba(74, 47, 189, 0.5) 111.58%); */
+  background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
   font-size: 20px;
   margin-bottom: 16px;
   display: inline-block;
@@ -389,7 +391,9 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
     position: absolute;
     width: 0;
     height: 100%;
-    background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%);
+    background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
+    /* background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%); */
+    /* background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%); */
     top: 0;
     left: 0;
     z-index: -1;
@@ -421,7 +425,9 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
 }
 .proj-imgbx::before {
   content: "";
-  background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%);
+  /* background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%); */
+  /* background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%); */
+  background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
   opacity: 0.85;
   position: absolute;
   width: 100%;
@@ -468,7 +474,9 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
 
 /************ Projects Css ************/
 .contact {
-  background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%);
+  /* background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%); */
+  /* background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%); */
+  background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
   padding: 60px 0 200px 0;
 }
 .contact img {
@@ -570,7 +578,9 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
 }
 .new-email-bx::before {
   content: "";
-  background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%);
+  /* background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%); */
+  /* background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%); */
+  background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
   border-radius: 20px;
   position: absolute;
   z-index: -1;
@@ -599,7 +609,9 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
   padding: 0 15px;
 }
 .new-email-bx button {
-  background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%);
+  /* background: linear-gradient(90.21deg, #AA367C -5.91%, #4A2FBD 111.58%); */
+  /* background: linear-gradient(90.21deg, #092327 -5.91%, #315b5b 111.58%); */
+  background: linear-gradient(90.21deg, #072333 -5.91%, #153241 111.58%);
   padding: 20px 65px;
   color: #fff;
   font-weight: 500;
